Revert "Change DssGrid to float to match DSS native 32-bit storage"
This reverts commit ba9fc64f3c5f8cc71438b6104aa5f6d72ad11550.

+++ src/main/java/mil/army/usace/hec/dss/DssGrid.java
@@ -17,10 +17,11 @@ import java.util.Objects;
  * <p>Cell coordinates are derived from the geometry — use {@link #x(int)} and {@link #y(int)}
  * to get the center coordinate of any cell in O(1).
  *
- * <p>Missing cell values are represented as {@link Float#NaN}.
+ * <p>Missing cell values are represented as {@link Double#NaN}.
  *
- * <p>Cell data uses {@code float} to match DSS native storage (32-bit floats),
- * ensuring lossless round-trips.
+ * <p><b>Precision note:</b> DSS stores grid data as 32-bit floats. This API uses
+ * doubles for convenience, but values will lose precision beyond ~7 significant
+ * digits when written and read back.
  *
  * <p>Create grids from scratch:
  * <pre>{@code
@@ -34,19 +35,19 @@ import java.util.Objects;
  * }</pre>
  */
 public final class DssGrid {
-    private final float[] data;
+    private final double[] data;
     private final int width;
     private final int height;
-    private final float cellSize;
-    private final float xOrigin;
-    private final float yOrigin;
+    private final double cellSize;
+    private final double xOrigin;
+    private final double yOrigin;
     private final String units;
     private final Crs crs;
     private final GridDataType dataType;
     private final NativeGridMetadata nativeMetadata;
 
-    private DssGrid(float[] data, int width, int height,
-                    float cellSize, float xOrigin, float yOrigin,
+    private DssGrid(double[] data, int width, int height,
+                    double cellSize, double xOrigin, double yOrigin,
                     String units, Crs crs, GridDataType dataType,
                     NativeGridMetadata nativeMetadata) {
         this.data = Objects.requireNonNull(data).clone();
@@ -87,8 +88,8 @@ public final class DssGrid {
      * @param crs      coordinate reference system
      * @param dataType what the cell values represent over time
      */
-    public static DssGrid of(float[] data, int width, int height,
-                             float cellSize, float xOrigin, float yOrigin,
+    public static DssGrid of(double[] data, int width, int height,
+                             double cellSize, double xOrigin, double yOrigin,
                              String units, Crs crs, GridDataType dataType) {
         return new DssGrid(data, width, height, cellSize, xOrigin, yOrigin,
                 units, crs, dataType, null);
@@ -98,7 +99,7 @@ public final class DssGrid {
      * Returns a new grid with the same geometry and CRS but different data, units, and data type.
      * This is the common case: applying model results to an existing grid's spatial layout.
      */
-    public DssGrid withData(float[] newData, String units, GridDataType dataType) {
+    public DssGrid withData(double[] newData, String units, GridDataType dataType) {
         return new DssGrid(newData, width, height, cellSize, xOrigin, yOrigin,
                 units, crs, dataType, null);
     }
@@ -122,10 +123,10 @@ public final class DssGrid {
     // ---- Data access ----
 
     /** Single cell value. Row 0 = north. */
-    public float value(int row, int col) { return data[row * width + col]; }
+    public double value(int row, int col) { return data[row * width + col]; }
 
     /** Flat copy of cell data, row-major, row 0 = north. */
-    public float[] data() { return data.clone(); }
+    public double[] data() { return data.clone(); }
 
     // ---- Geometry ----
 
@@ -136,19 +137,19 @@ public final class DssGrid {
     public int height() { return height; }
 
     /** Cell spacing in CRS units. */
-    public float cellSize() { return cellSize; }
+    public double cellSize() { return cellSize; }
 
     /** West edge of the grid. */
-    public float xOrigin() { return xOrigin; }
+    public double xOrigin() { return xOrigin; }
 
     /** South edge of the grid. */
-    public float yOrigin() { return yOrigin; }
+    public double yOrigin() { return yOrigin; }
 
     /** Center x-coordinate of the given column. */
-    public float x(int col) { return xOrigin + (col + 0.5f) * cellSize; }
+    public double x(int col) { return xOrigin + (col + 0.5) * cellSize; }
 
     /** Center y-coordinate of the given row. Row 0 = north. */
-    public float y(int row) { return yOrigin + (height - 1 - row + 0.5f) * cellSize; }
+    public double y(int row) { return yOrigin + (height - 1 - row + 0.5) * cellSize; }
 
     // ---- Semantics ----
 
@@ -167,8 +168,8 @@ public final class DssGrid {
     public NativeGridMetadata nativeMetadata() { return nativeMetadata; }
 
     /** @hidden */
-    public static DssGrid fromNative(float[] data, int width, int height,
-                                     float cellSize, float xOrigin, float yOrigin,
+    public static DssGrid fromNative(double[] data, int width, int height,
+                                     double cellSize, double xOrigin, double yOrigin,
                                      String units, Crs crs, GridDataType dataType,
                                      NativeGridMetadata nativeMetadata) {
         return new DssGrid(data, width, height, cellSize, xOrigin, yOrigin,
